ad_board_test/app/requirements.txt
Django==4.1.3
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-environ==0.11.2
drf-yasg==1.21.7
django-filter==23.2
